Support to force refresh index

+++ azure-pipelines/scripts/mirror/publish-debian-snapshot.sh
@@ -171,7 +171,7 @@ update_mirrors()
         elapsedseconds=$(($NOW_IN_SECONDS - $timestamp))
       fi
       # Refresh the index if more than 30 days (2592000 seconds), make sure the old indexes can be removed safely
-      if [ "$cursha256" == "$snsha256" ] && [ "$elapsedseconds" -lt 2592000 ]; then
+      if [ "$cursha256" == "$snsha256" ] && [ "$elapsedseconds" -lt 2592000 ] && [ "$FORCE_REFRESH" != "y" ]; then
         sudo ln -s ../../$dist_snapshot/dists/$dist $SNAPSHOT_TMP/dists/$dist
       else
         sudo cp -r $DISTS/$dist $SNAPSHOT_TMP/dists/
